Moved helper function to new project structure
